+++ Sources/VCore/Views and ViewControllers/CapsuleUIView.swift
@@ -11,6 +11,8 @@ import UIKit
 
 // MARK: - Capsule View
 /// `UIView` that rounds corners to capsule.
+///
+/// If width is greater than height, half of height will be taken as corner radius. If not, otherwise.
 open class CapsuleUIView: UIView {
     // MARK: Lifecycle
     open override func layoutSubviews() {
@@ -19,7 +21,13 @@ open class CapsuleUIView: UIView {
         // `clipsToBounds` and `layer.maskedCorners` are not necessary.
         // Plus, they conflict with shadows.
         
-        layer.cornerRadius = frame.height / 2
+        layer.cornerRadius = {
+            if frame.size.width > frame.size.height {
+                return frame.height / 2
+            } else {
+                return frame.width / 2
+            }
+        }()
     }
 }